feat: 更新的时间钩子

+++ packages/gin_example/models/models.go
@@ -19,8 +19,8 @@ var sqlDB *sql.DB
 
 type Model struct {
 	ID         int `gorm:"primary_key" json:"id"`
-	CreatedOn  int `json:"created_on"`
-	ModifiedOn int `json:"modified_on"`
+	CreatedOn  int64 `json:"created_on"`
+	ModifiedOn int64 `json:"modified_on"`
 }
 
 func init() {
